requirements.txt
aiocache==0.12.2
aiofiles==23.2.1
aiogram==3.5.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
beanie==1.26.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
curl_cffi==0.6.3
dateparser==1.2.0
dnspython==2.6.1
envparse==0.2.0
frozenlist==1.4.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.1.0
idna==3.7
lazy-model==0.2.0
loguru==0.7.2
magic-filter==1.0.12
motor==3.4.0
multidict==6.0.5
oauthlib==3.2.2
pyaes==1.6.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.7.1
PySocks==1.7.1
python-binance==1.0.19
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.4
regex==2024.4.28
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
StrEnum==0.4.15
Telethon==1.35.0
tls-client==1.0.1
toml==0.10.2
typing_extensions==4.11.0
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.1
websockets==12.0
yarl==1.9.4
APScheduler==3.10.4
